Remove `node-saml` code and use a local import for testing
