+++ package.json
@@ -51,6 +51,7 @@
   },
   "dependencies": {
     "debug": "^4.3.1",
+    "node-saml": "file:../node-saml",
     "passport-strategy": "^1.0.0",
     "xml-crypto": "^2.1.2",
     "xml-encryption": "^1.2.4",
