+++ src/index.ts
@@ -1,0 +1,15 @@
+import { Strategy, AbstractStrategy } from "./strategy";
+import { MultiSamlStrategy } from "./multiSamlStrategy";
+
+import type { VerifiedCallback, VerifyWithRequest, VerifyWithoutRequest } from "./types";
+
+export * from "node-saml";
+
+export {
+  AbstractStrategy,
+  Strategy,
+  MultiSamlStrategy,
+  VerifiedCallback,
+  VerifyWithRequest,
+  VerifyWithoutRequest,
+};
