--- src/node-saml/index.ts
@@ -1,4 +1,0 @@
-import type { CacheItem, CacheProvider } from "./inmemory-cache-provider";
-import { SAML } from "./saml";
-
-export { SAML, CacheItem, CacheProvider };
