--- src/node-saml/types.ts
@@ -1,126 +1,0 @@
-import type { CacheProvider } from "./inmemory-cache-provider";
-
-export type SignatureAlgorithm = "sha1" | "sha256" | "sha512";
-
-export interface SamlSigningOptions {
-  privateKey: string | Buffer;
-  signatureAlgorithm?: SignatureAlgorithm;
-  xmlSignatureTransforms?: string[];
-  digestAlgorithm?: string;
-}
-
-export const isValidSamlSigningOptions = (
-  options: Partial<SamlSigningOptions>
-): options is SamlSigningOptions => {
-  return options.privateKey != null;
-};
-
-export interface AudienceRestrictionXML {
-  Audience?: XMLObject[];
-}
-
-export type XMLValue = string | number | boolean | null | XMLObject | XMLValue[];
-
-export type XMLObject = {
-  [key: string]: XMLValue;
-};
-
-export type XMLInput = XMLObject;
-
-export type XMLOutput = Record<string, any>;
-
-export interface AuthorizeRequestXML {
-  "samlp:AuthnRequest": XMLInput;
-}
-
-export type CertCallback = (
-  callback: (err: Error | null, cert?: string | string[]) => void
-) => void;
-
-/**
- * These are SAML options that must be provided to construct a new SAML Strategy
- */
-export interface MandatorySamlOptions {
-  cert: string | string[] | CertCallback;
-}
-
-export interface SamlIDPListConfig {
-  entries: SamlIDPEntryConfig[];
-  getComplete?: string;
-}
-
-export interface SamlIDPEntryConfig {
-  providerId: string;
-  name?: string;
-  loc?: string;
-}
-
-export interface LogoutRequestXML {
-  "samlp:LogoutRequest": {
-    "saml:NameID": XMLInput;
-    [key: string]: XMLValue;
-  };
-}
-
-export interface ServiceMetadataXML {
-  EntityDescriptor: {
-    [key: string]: XMLValue;
-    SPSSODescriptor: XMLObject;
-  };
-}
-
-export type RacComparision = "exact" | "minimum" | "maximum" | "better";
-
-interface SamlScopingConfig {
-  idpList?: SamlIDPListConfig[];
-  proxyCount?: number;
-  requesterId?: string[] | string;
-}
-
-/**
- * The options required to use a SAML strategy
- * These may be provided by means of defaults specified in the constructor
- */
-export interface SamlOptions extends Partial<SamlSigningOptions>, MandatorySamlOptions {
-  // Core
-  callbackUrl?: string;
-  path: string;
-  protocol?: string;
-  host: string;
-  entryPoint?: string;
-  issuer: string;
-  decryptionPvk?: string | Buffer;
-
-  // Additional SAML behaviors
-  additionalParams: Record<string, string>;
-  additionalAuthorizeParams: Record<string, string>;
-  identifierFormat?: string | null;
-  acceptedClockSkewMs: number;
-  attributeConsumingServiceIndex?: string;
-  disableRequestedAuthnContext: boolean;
-  authnContext: string[];
-  forceAuthn: boolean;
-  skipRequestCompression: boolean;
-  authnRequestBinding?: string;
-  racComparison: RacComparision;
-  providerName?: string;
-  passive: boolean;
-  idpIssuer?: string;
-  audience?: string;
-  scoping?: SamlScopingConfig;
-  wantAssertionsSigned?: boolean;
-  maxAssertionAgeMs: number;
-
-  // InResponseTo Validation
-  validateInResponseTo: boolean;
-  requestIdExpirationPeriodMs: number;
-  cacheProvider: CacheProvider;
-
-  // Logout
-  logoutUrl: string;
-  additionalLogoutParams: Record<string, string>;
-  logoutCallbackUrl?: string;
-
-  // extras
-  disableRequestAcsUrl: boolean;
-}
