--- src/passport-saml/index.ts
@@ -1,26 +1,0 @@
-import type { CacheItem, CacheProvider } from "../node-saml/inmemory-cache-provider";
-import { SAML } from "../node-saml";
-import { Strategy, AbstractStrategy } from "./strategy";
-import { MultiSamlStrategy } from "./multiSamlStrategy";
-
-import type {
-  Profile,
-  SamlConfig,
-  VerifiedCallback,
-  VerifyWithRequest,
-  VerifyWithoutRequest,
-} from "./types";
-
-export {
-  SAML,
-  AbstractStrategy,
-  Strategy,
-  MultiSamlStrategy,
-  CacheItem,
-  CacheProvider,
-  Profile,
-  SamlConfig,
-  VerifiedCallback,
-  VerifyWithRequest,
-  VerifyWithoutRequest,
-};
